Add support for OG Images in the `/assets` folder
Previously, the OG Image was required to live next to the blog post and
be included as an image in the blog. This was so the Astro build could
run the image optimization. Sometimes the OG Image is not in the post,
and therefore could be added to the `/assets` folder and referenced
directly.

This change-set supports that.