Add a "Projects" section
Adds a section on the main page for "Projects". Posts will show up here
if they have the `featured` flag set to `true`. Projects can have a logo
and they will be rendered with the logo, name and description. Projects
will likely not have a publish date.

J2V8 and the AI Summary plugin are added as projects along with logos.