Add a Today I Learned section to the website
The Today I Learned (TIL) section will be a collection of short notes on
things I've learned. The notes will be tagged and searchable. The notes
are integrated with Obsidian, so I can write notes in markdown and
publish them to the website.

The TIL section will be paginated and have a dedicated page for each
note.

This change-set also fixes a few small issues with the website, such as
removing unused imports / variables, and handling some date errors.